reorder cfgs
